Move various test modules into `tests`
(with some modifications)
