+++ tests/physics/test_rotation.py
@@ -2,20 +2,22 @@ from panda3d.core import NodePath
 from panda3d.physics import (
     ActorNode,
     AngularEulerIntegrator,
+    AngularVectorForce,
     ForceNode,
     LinearEulerIntegrator,
     LinearFrictionForce,
     LinearVectorForce,
 )
+from direct.showbase.ShowBase import ShowBase
 
 
-class FallTest(NodePath):
+class RotationTest(NodePath):
     def __init__(self):
-        NodePath.__init__(self, "FallTest")
+        NodePath.__init__(self, "RotationTest")
 
     def setup(self):
         # Connect to Physics Manager:
-        self.actorNode = ActorNode("FallTestActorNode")
+        self.actorNode = ActorNode("RotationTestActorNode")
         #self.actorNode.getPhysicsObject().setOriented(1)
         #self.actorNode.getPhysical(0).setViscosity(0.1)
 
@@ -26,9 +28,9 @@ class FallTest(NodePath):
         avatarNodePath = base.loader.loadModel("models/misc/smiley")
         assert not avatarNodePath.isEmpty()
 
-        camLL = base.render.find("**/camLL")
-        camLL.reparentTo(avatarNodePath)
-        camLL.setPosHpr(0, -10, 0, 0, 0, 0)
+        # camLL = base.render.find("**/camLL")
+        # camLL.reparentTo(avatarNodePath)
+        # camLL.setPosHpr(0, -10, 0, 0, 0, 0)
         avatarNodePath.reparentTo(actorNodePath)
         #avatarNodePath.setPos(Vec3(0))
         #avatarNodePath.setHpr(Vec3(0))
@@ -36,19 +38,29 @@ class FallTest(NodePath):
         #self.phys = PhysicsManager()
         self.phys = base.physicsMgr
 
-        if 1:
-            fn=ForceNode("FallTest gravity")
-            fnp=NodePath(fn)
+        if 0:
+            fn = ForceNode("RotationTest gravity")
+            fnp = NodePath(fn)
             fnp.reparentTo(self)
             fnp.reparentTo(base.render)
-            gravity=LinearVectorForce(0.0, 0.0, -.5)
+            gravity=LinearVectorForce(0.0, 0.0, -0.5)
             fn.addForce(gravity)
             self.phys.addLinearForce(gravity)
             self.gravity = gravity
 
+        if 1:
+            fn = ForceNode("RotationTest spin")
+            fnp = NodePath(fn)
+            fnp.reparentTo(self)
+            fnp.reparentTo(base.render)
+            spin = AngularVectorForce(0.0, 0.0, 0.5)
+            fn.addForce(spin)
+            self.phys.addAngularForce(spin)
+            self.spin = spin
+
         if 0:
-            fn=ForceNode("FallTest viscosity")
-            fnp=NodePath(fn)
+            fn = ForceNode("RotationTest viscosity")
+            fnp = NodePath(fn)
             fnp.reparentTo(self)
             fnp.reparentTo(base.render)
             self.avatarViscosity=LinearFrictionForce(0.0, 1.0, 0)
@@ -64,17 +76,17 @@ class FallTest(NodePath):
         self.phys.attachPhysicalNode(self.actorNode)
 
         if 0:
-            self.momentumForce=LinearVectorForce(0.0, 0.0, 0.0)
-            fn=ForceNode("FallTest momentum")
-            fnp=NodePath(fn)
+            self.momentumForce = LinearVectorForce(0.0, 0.0, 0.0)
+            fn = ForceNode("RotationTest momentum")
+            fnp = NodePath(fn)
             fnp.reparentTo(base.render)
             fn.addForce(self.momentumForce)
             self.phys.addLinearForce(self.momentumForce)
 
         if 0:
-            self.acForce=LinearVectorForce(0.0, 0.0, 0.0)
-            fn=ForceNode("FallTest avatarControls")
-            fnp=NodePath(fn)
+            self.acForce = LinearVectorForce(0.0, 0.0, 0.0)
+            fn = ForceNode("RotationTest avatarControls")
+            fnp = NodePath(fn)
             fnp.reparentTo(base.render)
             fn.addForce(self.acForce)
             self.phys.addLinearForce(self.acForce)
@@ -86,10 +98,15 @@ class FallTest(NodePath):
         #self.actorNode.getPhysicsObject().resetPosition(self.avatarNodePath.getPos())
         #self.actorNode.updateTransform()
 
-if __name__ == "__main__":
-    from direct.directbase.ThreeUpStart import base
-    test = FallTest()
-    test.reparentTo(base.render)
-    test.setup()
-    base.camera.setY(-10.0)
-    base.run()
+
+def test_RotationTest():
+    base = ShowBase(windowType='none')
+    base.enableParticles()
+    base.addAngularIntegrator()
+    try:
+        test = RotationTest()
+        test.reparentTo(base.render)
+        test.setup()
+        # base.camera.setY(-10.0)
+    finally:
+        base.destroy()
